+++ packages/@expo/cli/src/export/exportApp.ts
@@ -55,7 +55,7 @@ export async function exportAppAsync(
   const exp = await getPublicExpoManifestAsync(projectRoot);
 
   const useWebSSG = exp.web?.output === 'static';
-  const assetPrefix = exp.assetPrefix?.replace(/\/+$/, '') ?? '';
+  const basePath = exp.experiments?.basePath?.replace(/\/+$/, '') ?? '';
   const publicPath = path.resolve(projectRoot, env.EXPO_PUBLIC_FOLDER);
 
   const outputPath = path.resolve(projectRoot, outputDir);
@@ -119,23 +119,23 @@ export async function exportAppAsync(
         outputDir: outputPath,
         // TODO: Expose
         minify,
-        assetPrefix,
+        basePath: basePath,
       });
       Log.log('Finished saving static files');
     } else {
       const cssLinks = await exportCssAssetsAsync({
         outputDir,
         bundles,
-        assetPrefix,
+        basePath: basePath,
       });
       let html = await createTemplateHtmlFromExpoConfigAsync(projectRoot, {
-        scripts: [`${assetPrefix}/bundles/${fileNames.web}`],
+        scripts: [`${basePath}/bundles/${fileNames.web}`],
         cssLinks,
       });
       // Add the favicon assets to the HTML.
       const modifyHtml = await getVirtualFaviconAssetsAsync(projectRoot, {
         outputDir,
-        assetPrefix,
+        basePath: basePath,
       });
       if (modifyHtml) {
         html = modifyHtml(html);
@@ -152,7 +152,7 @@ export async function exportAppAsync(
       await persistMetroAssetsAsync(bundles.web.assets, {
         platform: 'web',
         outputDirectory: staticFolder,
-        assetPrefix,
+        basePath: basePath,
       });
     }
   }