+++ packages/@expo/cli/src/export/exportAssets.ts
@@ -108,11 +108,11 @@ export async function exportAssetsAsync(
 export async function exportCssAssetsAsync({
   outputDir,
   bundles,
-  assetPrefix,
+  basePath,
 }: {
   bundles: Partial<Record<ModPlatform, BundleOutput>>;
   outputDir: string;
-  assetPrefix: string;
+  basePath: string;
 }) {
   const assets = uniqBy(
     Object.values(bundles).flatMap((bundle) => bundle!.css),
@@ -128,5 +128,5 @@ export async function exportCssAssetsAsync({
     assets.map((v) => fs.promises.writeFile(path.join(outputDir, v.filename), v.source))
   );
 
-  return assets.map((v) => assetPrefix + '/' + v.filename);
+  return assets.map((v) => basePath + '/' + v.filename);
 }